making a site more accessible
Websites should be designed in a way that allows everyone to access and interact with their content easily, regardless of any disabilities they may have. This course on web accessibility has taught me the importance of considering the needs of all individuals when designing a website. By making websites inclusive and user-friendly, we can ensure that information is democratized and available to everyone.